requirements.txt
git+https://github.com/datacommonsorg/llm-tools
llama-cpp-python
huggingface-hub
rich
langchain-anthropic >= 0.1.19